requirements.txt
google-generativeai
PyYAML
feedparser